+++ packages/core/src/js/index.js
@@ -20,5 +20,6 @@ export { installHovercard } from './hovercard.js';
 export { installToggleGroup } from './toggle-group.js';
 export { installContextMenu } from './context-menu.js';
 export { installCommand } from './command.js';
+export { installCalendar } from './calendar.js';
 
 export const version = '0.0.0';